Use correct type to check for errors

+++ src/process/handle.c
@@ -28,7 +28,7 @@ bool process_handle_is_valid(ProcessHandle* handle) {
 }
 
 size_t process_handle_read(ProcessHandle* handle, MemoryAddress addr, void* buf, size_t size) {
-    uint64_t did_read = pread(handle->mem, buf, size, addr);
+    ssize_t did_read = pread(handle->mem, buf, size, addr);
     if(did_read < 0) {
         perror("Error reading process memory");
         return 0;
@@ -37,7 +37,7 @@ size_t process_handle_read(ProcessHandle* handle, MemoryAddress addr, void* buf,
 }
 
 size_t process_handle_write(ProcessHandle* handle, MemoryAddress addr, void* buf, size_t size) {
-    uint64_t did_write = pwrite(handle->mem, buf, size, addr);
+    ssize_t did_write = pwrite(handle->mem, buf, size, addr);
     if(did_write < 0) {
         perror("Error writing process memory");
         return 0;
